Add stores query relocation
