+++ packages/system/src/Store/api/operations.graphql
@@ -1,0 +1,12 @@
+query stores($first: Int, $offset: Int, $filter: StoreFilterInput) {
+  stores(page: { first: $first, offset: $offset }, filter: $filter) {
+    ... on StoreConnector {
+      __typename
+      totalCount
+      nodes {
+        code
+        id
+      }
+    }
+  }
+}
