--- packages/system/src/Store/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './useStores';
