+++ packages/system/src/Store/index.ts
@@ -1,1 +1,1 @@
-export * from './hooks';
+export * from './api';
